+++ packages/bottender/src/server/Server.ts
@@ -4,12 +4,13 @@ import { IncomingMessage, ServerResponse } from 'http';
 
 import fromEntries from 'fromentries';
 import merge from 'lodash/merge';
+import { match } from 'path-to-regexp';
 import { pascalcase } from 'messaging-api-common';
 
 import Bot from '../bot/Bot';
 import getBottenderConfig from '../shared/getBottenderConfig';
 import getSessionStore from '../getSessionStore';
-import { Action, BottenderConfig, Plugin } from '../types';
+import { Action, BottenderConfig, Plugin, RequestContext } from '../types';
 
 export type ServerOptions = {
   useConsole?: boolean;
@@ -130,15 +131,24 @@ class Server {
     for (let i = 0; i < this._channelBots.length; i++) {
       const { webhookPath, bot } = this._channelBots[i];
 
-      if (pathname === webhookPath) {
-        const result = (bot.connector as any).preprocess({
-          method: req.method,
-          url: `https://${req.headers.host}${req.url}`,
-          headers: req.headers,
+      const matchPath = match<Record<string, string>>(webhookPath);
+      const matchResult = matchPath(pathname);
+
+      if (matchResult) {
+        const httpContext: RequestContext = {
+          id: (req as any).id,
+          method: req.method as string,
+          path: pathname,
           query,
+          headers: req.headers,
           rawBody: (req as any).rawBody,
           body: (req as any).body,
-        });
+          params: matchResult.params,
+          url: `https://${req.headers.host}${req.url}`,
+        };
+
+        // eslint-disable-next-line no-await-in-loop
+        const result = await (bot.connector as any).preprocess(httpContext);
 
         const { shouldNext } = result;
         let { response } = result;
@@ -174,12 +184,7 @@ class Server {
             ...query,
             ...(req as any).body,
           },
-          {
-            method: req.method as string,
-            path: pathname,
-            query,
-            headers: req.headers as Record<string, string>,
-          }
+          httpContext
         );
 
         if (response) {