+++ packages/bottender/src/slack/__tests__/SlackConnector.spec.ts
@@ -445,6 +445,15 @@ describe('#mapRequestToEvents', () => {
     expect(events).toHaveLength(1);
     expect(events[0]).toBeInstanceOf(SlackEvent);
   });
+
+  it('should include callbackId when request is a interactiveMessageRequest', () => {
+    const { connector } = setup();
+    const events = connector.mapRequestToEvents(interactiveMessageRequest.body);
+
+    expect(events).toHaveLength(1);
+    expect(events[0]).toBeInstanceOf(SlackEvent);
+    expect(events[0].callbackId).toBe('wopr_game');
+  });
 });
 
 describe('#createContext', () => {