+++ packages/bottender/src/telegram/TelegramBot.ts
@@ -1,9 +1,12 @@
 import { TelegramClient } from 'messaging-api-telegram';
 
-import Bot from '../bot/Bot';
+import Bot, { OnRequest } from '../bot/Bot';
 import SessionStore from '../session/SessionStore';
 
-import TelegramConnector, { TelegramRequestBody } from './TelegramConnector';
+import TelegramConnector, {
+  TelegramConnectorOptions,
+  TelegramRequestBody,
+} from './TelegramConnector';
 import TelegramContext from './TelegramContext';
 import TelegramEvent from './TelegramEvent';
 
@@ -11,7 +14,7 @@ type PollingOptions = {
   offset?: number;
   limit?: number;
   timeout?: number;
-  allowed_updates?: string[];
+  allowedUpdates?: string[];
 };
 
 export default class TelegramBot extends Bot<
@@ -25,18 +28,17 @@ export default class TelegramBot extends Bot<
   _shouldGetUpdates: boolean;
 
   constructor({
-    accessToken,
     sessionStore,
     sync,
-    origin,
-  }: {
-    accessToken: string;
+    onRequest,
+    ...connectorOptions
+  }: TelegramConnectorOptions & {
     sessionStore?: SessionStore;
     sync?: boolean;
-    origin?: string;
+    onRequest?: OnRequest;
   }) {
-    const connector = new TelegramConnector({ accessToken, origin });
-    super({ connector, sessionStore, sync });
+    const connector = new TelegramConnector(connectorOptions);
+    super({ connector, sessionStore, sync, onRequest });
 
     this._offset = null;
     this._shouldGetUpdates = false;
@@ -69,7 +71,7 @@ export default class TelegramBot extends Bot<
           }
 
           const highestUpdateId = Math.max(
-            ...updates.map((update: any) => update.update_id)
+            ...updates.map((update: any) => update.updateId)
           );
 
           this._offset = highestUpdateId + 1;