+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.4.5 / 2020-05-11
+
+- [fix] fix issue #618 ngrok undefined error message (#765)
+
+# 1.4.4 / 2020-05-06
+
+## slack
+
+- [fix] convert slack interactive message event to camelcase (#755).
+
 # 1.4.3 / 2020-04-29
 
 - [type] use string instead enum to compare.